lambda/packages/transactions_lambda_package/requirements.txt
fastapi==0.135.1
mangum==0.21.0
pydantic==2.12.5
python-dotenv==1.2.2
numpy==2.2.6
onnxruntime==1.24.3